Abstract OmhTask
